Redesign asynchronous command interface
The worker thread system is now a little less typesafe interanally: Engine::sendToWorker and Engine::sendFromWorker have to be called from the corresponding context. The plugin API on the other hand is now more typesafe: An interface struct corresponding to the current context is passed to the command perform callback, making it harder to accidentally calling API functions from the wrong context.