Don't set number of channels from audio session
Don't override the number of channels passed in the audio driver
options with the values obtained from the audio session. The number
of channels can change when the route changes (e.g. speaker vs
headphones) and we usually want a fixed number of channels in the audio
unit (or a way to be notified of changes in the number of channels).